requirements.txt
pytube >=12.1.3, <12.2
moviepy >= 1.0.3, <1.1
openai-whisper